size

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(void)
 {
-	printf("%d\n Size of int: ", sizeof(int));
+	printf("%d\n Size of int: ", sizeof(long int));
 	return (0);
 }
